Split tests (#2891)
