+++ Cargo.toml
@@ -16,7 +16,6 @@ members = [
   "crates/hotshot-stake-table",
   "crates/libp2p-networking",
   "crates/macros",
-  "crates/testing-macros",
   "crates/task",
   "crates/task-impls",
   "crates/testing",
