+++ crates/testing/tests/tests_1.rs
@@ -1,0 +1,3 @@
+mod tests_1 {
+    automod::dir!("tests/tests_1");
+}
