+++ crates/testing/tests/tests_1/view_sync_task.rs
