+++ crates/testing/tests/tests_2.rs
@@ -1,0 +1,3 @@
+mod tests_2 {
+    automod::dir!("tests/tests_2");
+}
