+++ crates/testing/tests/tests_3.rs
@@ -1,0 +1,3 @@
+mod tests_3 {
+    automod::dir!("tests/tests_3");
+}
