+++ crates/testing/tests/tests_4.rs
@@ -1,0 +1,3 @@
+mod tests_4 {
+    automod::dir!("tests/tests_4");
+}
