+++ crates/testing/tests/tests_5.rs
@@ -1,0 +1,3 @@
+mod tests_5 {
+    automod::dir!("tests/tests_5");
+}
