--- crates/testing/tests/unit.rs
@@ -1,3 +1,0 @@
-mod unit {
-    mod message;
-}
